Program.Main zredukován na parsování args + volání enginu. Přidán FractalConfig.FromArgs a Polynomial.FromArgs.

+++ NNPTPZ1/Mathematics/Complex.cs
@@ -13,7 +13,7 @@ namespace NNPTPZ1.Mathematics
         public Complex(double real, double imaginary)
         {
             Real = real;
-            Imaginary = (float)imaginary;
+            Imaginary = imaginary;
         }
 
         public double Abs()
